requirements.txt
# Multi-Agent Debate System Requirements

# AutoGen for Multi-Agent Conversations (최신 버전)
pyautogen>=0.2.0
autogen-ext>=0.2.0

# LangChain for RAG (최신 안정 버전)
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# OpenAI
openai>=1.10.0

# Vector Store
chromadb>=0.4.0
faiss-cpu>=1.7.0

# Embeddings (선택사항 - HuggingFace 사용 시)
sentence-transformers>=2.2.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
tiktoken>=0.5.0

# UI
streamlit>=1.28.0
gradio>=4.0.0
plotly>=5.17.0
